Restructure pages (#287)
* rmv repeated code

* move files

* update pkgs

* dynamic pages created

* dynamic folder addn in src/apps/pages

* rename dir name

* format text
