+++ src/apps/pages/programs/API/wolframSolver.py
