+++ src/apps/pages/programs/Simple/cryptoCurrency.py
