+++ src/apps/pages/programs/Simple/joke.py
