+++ src/apps/pages/programs/Study/translator.py
